@@ -1,19 +1,15 @@
 <?php
 
+/*
+ * This file is part of OAuth 2.0 Laravel.
+ *
+ * (c) Luca Degasperi <packages@lucadegasperi.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 return [
-
-    /*
-    |--------------------------------------------------------------------------
-    | Database Connection to use
-    |--------------------------------------------------------------------------
-    |
-    | Set the default database connection to use for the repositories, when
-    | set to default, it uses whatever connection you specified in your
-    | laravel database config.
-    |
-    */
-
-    'database' => 'default',
 
     /*
     |--------------------------------------------------------------------------
