@@ -1,26 +1,48 @@
 <?php
-/**
- * Fluent storage implementation for an OAuth 2.0 Scope
+
+/*
+ * This file is part of OAuth 2.0 Laravel.
  *
- * @package   lucadegasperi/oauth2-server-laravel
- * @author    Luca Degasperi <luca@lucadegasperi.com>
- * @copyright Copyright (c) Luca Degasperi
- * @licence   http://mit-license.org/
- * @link      https://github.com/lucadegasperi/oauth2-server-laravel
+ * (c) Luca Degasperi <packages@lucadegasperi.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
  */
 
 namespace LucaDegasperi\OAuth2Server\Storage;
 
 use Illuminate\Database\ConnectionResolverInterface as Resolver;
-use League\OAuth2\Server\Storage\ScopeInterface;
 use League\OAuth2\Server\Entity\ScopeEntity;
+use League\OAuth2\Server\Storage\ScopeInterface;
 
-class FluentScope extends FluentAdapter implements ScopeInterface
+/**
+ * This is the fluent scope class.
+ *
+ * @author Luca Degasperi <packages@lucadegasperi.com>
+ */
+class FluentScope extends AbstractFluentAdapter implements ScopeInterface
 {
+    /*
+     * Limit clients to scopes.
+     *
+     * @var bool
+     */
     protected $limitClientsToScopes = false;
 
+    /*
+     * Limit scopes to grants.
+     *
+     * @var bool
+     */
     protected $limitScopesToGrants = false;
 
+    /**
+     * Create a new fluent scope instance.
+     *
+     * @param \Illuminate\Database\ConnectionResolverInterface $resolver
+     * @param bool|false $limitClientsToScopes
+     * @param bool|false $limitScopesToGrants
+     */
     public function __construct(Resolver $resolver, $limitClientsToScopes = false, $limitScopesToGrants = false)
     {
         parent::__construct($resolver);
@@ -28,29 +50,48 @@ class FluentScope extends FluentAdapter implements ScopeInterface
         $this->limitScopesToGrants = $limitScopesToGrants;
     }
 
+    /**
+     * Set limit clients to scopes.
+     *
+     * @param bool|false $limit
+     */
     public function limitClientsToScopes($limit = false)
     {
         $this->limitClientsToScopes = $limit;
     }
 
+    /**
+     * Set limit scopes to grants.
+     *
+     * @param bool|false $limit
+     */
     public function limitScopesToGrants($limit = false)
     {
         $this->limitScopesToGrants = $limit;
     }
 
+    /**
+     * Check if clients are limited to scopes.
+     *
+     * @return bool|false
+     */
     public function areClientsLimitedToScopes()
     {
         return $this->limitClientsToScopes;
     }
 
+    /**
+     * Check if scopes are limited to grants.
+     *
+     * @return bool|false
+     */
     public function areScopesLimitedToGrants()
     {
         return $this->limitScopesToGrants;
     }
 
-
     /**
-     * Return information about a scope
+     * Return information about a scope.
      *
      * Example SQL query:
      *
@@ -58,23 +99,24 @@ class FluentScope extends FluentAdapter implements ScopeInterface
      * SELECT * FROM oauth_scopes WHERE scope = :scope
      * </code>
      *
-     * @param  string     $scope     The scope
-     * @param  string     $grantType The grant type used in the request (default = "null")
-     * @param  string     $clientId  The client id used for the request (default = "null")
+     * @param string $scope The scope
+     * @param string $grantType The grant type used in the request (default = "null")
+     * @param string $clientId The client id used for the request (default = "null")
+     *
      * @return \League\OAuth2\Server\Entity\ScopeEntity|null If the scope doesn't exist return false
      */
     public function get($scope, $grantType = null, $clientId = null)
     {
-         $query = $this->getConnection()->table('oauth_scopes')
+        $query = $this->getConnection()->table('oauth_scopes')
                     ->select('oauth_scopes.id as id', 'oauth_scopes.description as description')
                     ->where('oauth_scopes.id', $scope);
 
-        if ($this->limitClientsToScopes === true and ! is_null($clientId)) {
+        if ($this->limitClientsToScopes === true && !is_null($clientId)) {
             $query = $query->join('oauth_client_scopes', 'oauth_scopes.id', '=', 'oauth_client_scopes.scope_id')
                            ->where('oauth_client_scopes.client_id', $clientId);
         }
 
-        if ($this->limitScopesToGrants === true and ! is_null($grantType)) {
+        if ($this->limitScopesToGrants === true && !is_null($grantType)) {
             $query = $query->join('oauth_grant_scopes', 'oauth_scopes.id', '=', 'oauth_grant_scopes.scope_id')
                            ->join('oauth_grants', 'oauth_grants.id', '=', 'oauth_grant_scopes.grant_id')
                            ->where('oauth_grants.id', $grantType);
@@ -83,14 +125,15 @@ class FluentScope extends FluentAdapter implements ScopeInterface
         $result = $query->first();
 
         if (is_null($result)) {
-            return null;
+            return;
         }
 
         $scope = new ScopeEntity($this->getServer());
         $scope->hydrate([
             'id' => $result->id,
-            'description' => $result->description
+            'description' => $result->description,
         ]);
+
         return $scope;
     }
 }
