@@ -1,27 +1,35 @@
 <?php
-/**
- * Fluent storage implementation for an OAuth 2.0 Session
+
+/*
+ * This file is part of OAuth 2.0 Laravel.
  *
- * @package   lucadegasperi/oauth2-server-laravel
- * @author    Luca Degasperi <luca@lucadegasperi.com>
- * @copyright Copyright (c) Luca Degasperi
- * @licence   http://mit-license.org/
- * @link      https://github.com/lucadegasperi/oauth2-server-laravel
+ * (c) Luca Degasperi <packages@lucadegasperi.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
  */
+
 namespace LucaDegasperi\OAuth2Server\Storage;
 
+use Carbon\Carbon;
 use League\OAuth2\Server\Entity\AccessTokenEntity;
 use League\OAuth2\Server\Entity\AuthCodeEntity;
-use League\OAuth2\Server\Storage\SessionInterface;
-use League\OAuth2\Server\Entity\SessionEntity;
 use League\OAuth2\Server\Entity\ScopeEntity;
-use Carbon\Carbon;
+use League\OAuth2\Server\Entity\SessionEntity;
+use League\OAuth2\Server\Storage\SessionInterface;
 
-class FluentSession extends FluentAdapter implements SessionInterface
+/**
+ * This is the fluent session class.
+ *
+ * @author Luca Degasperi <packages@lucadegasperi.com>
+ */
+class FluentSession extends AbstractFluentAdapter implements SessionInterface
 {
     /**
-     * Get a session from it's identifier
+     * Get a session from it's identifier.
+     *
      * @param string $sessionId
+     *
      * @return \League\OAuth2\Server\Entity\SessionEntity
      */
     public function get($sessionId)
@@ -30,8 +38,8 @@ class FluentSession extends FluentAdapter implements SessionInterface
                     ->where('oauth_sessions.id', $sessionId)
                     ->first();
 
-        if(is_null($result)) {
-            return null;
+        if (is_null($result)) {
+            return;
         }
 
         return (new SessionEntity($this->getServer()))
@@ -40,8 +48,10 @@ class FluentSession extends FluentAdapter implements SessionInterface
     }
 
     /**
-     * Get a session from an access token
-     * @param  \League\OAuth2\Server\Entity\AccessTokenEntity $accessToken The access token
+     * Get a session from an access token.
+     *
+     * @param \League\OAuth2\Server\Entity\AccessTokenEntity $accessToken The access token
+     *
      * @return \League\OAuth2\Server\Entity\SessionEntity
      */
     public function getByAccessToken(AccessTokenEntity $accessToken)
@@ -53,7 +63,7 @@ class FluentSession extends FluentAdapter implements SessionInterface
                 ->first();
 
         if (is_null($result)) {
-            return null;
+            return;
         }
 
         return (new SessionEntity($this->getServer()))
@@ -62,8 +72,10 @@ class FluentSession extends FluentAdapter implements SessionInterface
     }
 
     /**
-     * Get a session's scopes
-     * @param  \League\OAuth2\Server\Entity\SessionEntity
+     * Get a session's scopes.
+     *
+     * @param \League\OAuth2\Server\Entity\SessionEntity
+     *
      * @return array Array of \League\OAuth2\Server\Entity\ScopeEntity
      */
     public function getScopes(SessionEntity $session)
@@ -74,58 +86,64 @@ class FluentSession extends FluentAdapter implements SessionInterface
                   ->join('oauth_scopes', 'oauth_session_scopes.scope_id', '=', 'oauth_scopes.id')
                   ->where('oauth_session_scopes.session_id', $session->getId())
                   ->get();
-        
+
         $scopes = [];
-        
+
         foreach ($result as $scope) {
             $scopes[] = (new ScopeEntity($this->getServer()))->hydrate([
                 'id' => $scope->id,
                 'description' => $scope->description,
             ]);
         }
-        
+
         return $scopes;
     }
 
     /**
-     * Create a new session
-     * @param  string $ownerType         Session owner's type (user, client)
-     * @param  string $ownerId           Session owner's ID
-     * @param  string $clientId          Client ID
-     * @param  string $clientRedirectUri Client redirect URI (default = null)
-     * @return integer The session's ID
+     * Create a new session.
+     *
+     * @param string $ownerType Session owner's type (user, client)
+     * @param string $ownerId Session owner's ID
+     * @param string $clientId Client ID
+     * @param string $clientRedirectUri Client redirect URI (default = null)
+     *
+     * @return int The session's ID
      */
     public function create($ownerType, $ownerId, $clientId, $clientRedirectUri = null)
     {
         return $this->getConnection()->table('oauth_sessions')->insertGetId([
-            'client_id'  => $clientId,
+            'client_id' => $clientId,
             'owner_type' => $ownerType,
-            'owner_id'   => $ownerId,
+            'owner_id' => $ownerId,
             'client_redirect_uri' => $clientRedirectUri,
             'created_at' => Carbon::now(),
-            'updated_at' => Carbon::now()
+            'updated_at' => Carbon::now(),
         ]);
     }
 
     /**
-     * Associate a scope with a session
-     * @param  \League\OAuth2\Server\Entity\SessionEntity $session
-     * @param  \League\OAuth2\Server\Entity\ScopeEntity $scope The scopes ID might be an integer or string
+     * Associate a scope with a session.
+     *
+     * @param \League\OAuth2\Server\Entity\SessionEntity $session
+     * @param \League\OAuth2\Server\Entity\ScopeEntity $scope The scopes ID might be an integer or string
+     *
      * @return void
      */
     public function associateScope(SessionEntity $session, ScopeEntity $scope)
     {
         $this->getConnection()->table('oauth_session_scopes')->insert([
             'session_id' => $session->getId(),
-            'scope_id'   => $scope->getId(),
+            'scope_id' => $scope->getId(),
             'created_at' => Carbon::now(),
-            'updated_at' => Carbon::now()
+            'updated_at' => Carbon::now(),
         ]);
     }
 
     /**
-     * Get a session from an auth code
-     * @param  \League\OAuth2\Server\Entity\AuthCodeEntity $authCode The auth code
+     * Get a session from an auth code.
+     *
+     * @param \League\OAuth2\Server\Entity\AuthCodeEntity $authCode The auth code
+     *
      * @return \League\OAuth2\Server\Entity\SessionEntity
      */
     public function getByAuthCode(AuthCodeEntity $authCode)
@@ -137,7 +155,7 @@ class FluentSession extends FluentAdapter implements SessionInterface
             ->first();
 
         if (is_null($result)) {
-            return null;
+            return;
         }
 
         return (new SessionEntity($this->getServer()))
