@@ -1,9 +1,18 @@
 <?php
 
+/*
+ * This file is part of OAuth 2.0 Laravel.
+ *
+ * (c) Luca Degasperi <packages@lucadegasperi.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 use LucaDegasperi\OAuth2Server\Storage\FluentAccessToken;
 use Mockery as m;
 
-class FluentAccessTokenTest extends DBTestCase
+class FluentAccessTokenTest extends AbstractDBTestCase
 {
     public function getAccessTokenRepository()
     {
