@@ -1,9 +1,18 @@
 <?php
 
+/*
+ * This file is part of OAuth 2.0 Laravel.
+ *
+ * (c) Luca Degasperi <packages@lucadegasperi.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 use LucaDegasperi\OAuth2Server\Storage\FluentScope;
 use Mockery as m;
 
-class FluentScopeTest extends DBTestCase
+class FluentScopeTest extends AbstractDBTestCase
 {
     public function getScopeRepository()
     {
@@ -85,7 +94,7 @@ class FluentScopeTest extends DBTestCase
     public function test_get_scope()
     {
         $repo = $this->getScopeRepository();
-        $result = $repo->get('scope1');  
+        $result = $repo->get('scope1');
 
         $this->assertIsScope($result);
     }
